+++ Changelog.md
@@ -15,6 +15,7 @@
 - CS-385 - Created GameDevelopmentGoal model, seeder controller, backend list and form
 - CS-385 - Changed migrations, renamed "order" columns to "sort_order"
 - CS-363 - Fixed success/error message issue on back list bulk delete
+- CS-454 - Remove the Team column from the Mandates tables
 - CS-455 - Sort the teams in the menu
 - CS-370 - Create Knowledge Repository and Knowledge Repository Parameters menus on backend
 - CS-451 - Membership card cannot be activated if the assigned scout is inactive